+++ package.json
@@ -45,7 +45,7 @@
     "@ledgerhq/hw-transport": "6.27.1",
     "@ledgerhq/hw-transport-webhid": "6.27.1",
     "@ledgerhq/logs": "6.12.0",
-    "@mayanfinance/swap-sdk": "^10.9.3",
+    "@mayanfinance/swap-sdk": "10.9.3",
     "@metaplex-foundation/mpl-token-metadata": "3.4.0",
     "@metaplex-foundation/umi": "0.9.2",
     "@metaplex-foundation/umi-bundle-defaults": "0.9.2",
@@ -54,7 +54,7 @@
     "@reduxjs/toolkit": "^2.5.1",
     "@solana/spl-token": "0.4.0",
     "@solana/wallet-adapter-wallets": "0.19.32",
-    "@testnet-mayan/swap-sdk": "^1.2.0",
+    "@testnet-mayan/swap-sdk": "1.2.0",
     "@wormhole-foundation/sdk": "3.3.0",
     "@wormhole-foundation/sdk-definitions": "3.3.0",
     "@wormhole-foundation/sdk-definitions-ntt": "1.0.6",
