+++ src/exports/mayan.ts
@@ -1,6 +1,5 @@
 export {
   MayanRoute,
-  MayanRouteBase,
   MayanRouteMCTP,
   MayanRouteMONOCHAIN,
   MayanRouteSWIFT,
