+++ src/routes/mayan/index.ts
@@ -1,6 +1,5 @@
 export { default as MayanRoute } from './MayanRoute';
-export { default as MayanRouteBase } from './MayanRouteBase';
 export { default as MayanRouteMCTP } from './MayanRouteMCTP';
 export { default as MayanRouteMONOCHAIN } from './MayanRouteMONOCHAIN';
-export { default as MayanRouteSWIFT } from './MayanRouteSwift';
+export { default as MayanRouteSWIFT } from './MayanRouteSWIFT';
 export { default as MayanRouteWH } from './MayanRouteWH';
